requirements.txt
streamlit
pandas
gspread
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib